add up configuration
I explicitly removed the production domain. Because there is an AWS
Lambda behind this, I would rather people not programatically find this
and run up my bill. Yes, this is definitely a very paranoid thing to do
and the whole point of having a website is so that people see it.